+++ src/hip/handlehip.cpp
@@ -35,6 +35,7 @@
 #include <miopen/invoker.hpp>
 #include <miopen/kernel_cache.hpp>
 #include <miopen/logger.hpp>
+#include <miopen/timer.hpp>
 
 #if !MIOPEN_ENABLE_SQLITE_KERN_CACHE
 #include <miopen/write_file.hpp>
@@ -371,20 +372,11 @@ Program Handle::LoadProgram(const std::string& program_name,
         this->GetDeviceName(), this->GetMaxComputeUnits(), program_name, params, is_kernel_str);
     if(hsaco.empty())
     {
-
-#if MIOPEN_BUILD_DEV
-        auto compileStart = std::chrono::steady_clock::now();
-#endif
+	    
+        CompileTimer ct;
         auto p =
             HIPOCProgram{program_name, params, is_kernel_str, this->GetDeviceName(), kernel_src};
-
-#if MIOPEN_BUILD_DEV
-        auto compileEnd = std::chrono::steady_clock::now();
-        auto timems     = std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(
-                          compileEnd - compileStart)
-                          .count();
-        MIOPEN_LOG_I2("Kernel " << program_name << " Compile Time: " << timems << " ms");
-#endif
+        ct.Log("Kernel", program_name);
 
 // Save to cache
 #if MIOPEN_ENABLE_SQLITE_KERN_CACHE
